Restore compatibility with elasticsearch 2.x error logging
Between elasticsearch-py 2.x and 5.x, the signature of the log_request_fail
method on the Connection object changed.  An additional `path` parameter was
added.  As a result of this, there is an unexpected exception raised by
tornado_elasticsearch during the handling an HTTP error.  This commit adds a
check to determine the version of elasticsearch and uses the appropriate call
signature.